+++ apps/ssr-testing/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lodado/ssr-testing
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @lodado/react-namespace@1.0.7
+
 ## 0.0.7
 
 ### Patch Changes